Refactored code into separate classes.
