+++ client/src/GetDataFromDb.js
@@ -1,0 +1,7 @@
+const GetDataFromDbClass = require('./GetDataFromDb.class');
+let GetDataFromDb;
+
+module.exports = function() {
+    GetDataFromDb = GetDataFromDb || new GetDataFromDbClass();
+    return GetDataFromDb;
+};
