+++ client/src/ModifyDropdowns.js
@@ -1,0 +1,71 @@
+import React, { Component } from 'react';
+
+import axios from 'axios';
+
+/*
+Performs checks on the validity of the input to the database.
+*/
+class ModifyDropdowns {
+	
+	constructor() {
+	}
+	
+	    // This transforms the data object property temperature into an array!
+    // Source: https://medium.com/poka-techblog/simplify-your-javascript-use-map-reduce-and-filter-bd02c593cc2d
+    getArrayOfOneElementType(data) {
+        var officersIds = [];
+        data.forEach(function (data) {officersIds.push(data.temperature);})   
+        console.log(officersIds[2]);
+        return officersIds;
+    }
+    
+	// This transforms the data object property temperature into an array!
+    // Source: https://medium.com/poka-techblog/simplify-your-javascript-use-map-reduce-and-filter-bd02c593cc2d
+    getSingleEntry(data) {
+        var officersIds = [];
+        data.forEach(function (data) {officersIds.push(data.temperature);})   
+        console.log(officersIds[2]);
+        return officersIds[2];
+    }
+	
+	// This function adds an option to the dropdownbox.
+	addOptionToDropdown(){
+        var inputElemAdd = document.getElementsByTagName('select');
+        var selectBox = document.getElementById("dynamic-select");        
+        selectBox[0].label = "Wrote 0";
+        selectBox[2].label = "Wrote 2";
+        selectBox[3] = new Option('hi, added last label', 'id0',false,false); // add option
+    }
+	
+
+	sayHello(arr) {	
+		var inputElemAdd = document.getElementsByTagName('select');
+        var selectBox = document.getElementById("dynamic-select");        
+		var i
+		var newArr = this.filterUndefineds(arr);
+		
+		// set length of dropdownbox to incomming array
+		selectBox.length = newArr.length;
+		
+		// assign arrays
+		for (i = 0; i < newArr.length; i++) {
+			selectBox[i].label = newArr[i];
+		}
+	}
+	
+	// Filter undefined values from an array
+	filterUndefineds(arr){
+		var i
+		var newArr = [];
+		
+		// only copy the non-undefined values from incoming array to newArr
+		for (i = 0; i < arr.length; i++) {
+			if (arr[i] !== undefined) {
+				newArr.push(arr[i]);
+			}
+		}
+		return newArr;
+	}
+}
+
+//export default FormatChecks;
